:recycle: Merge constant and type to the root level
